migration to Spring 5.3.9.RELEASE

+++ src/modules/rest/auditing/src/main/java/it/geosolutions/geostore/services/rest/auditing/AuditInfoExtractor.java
@@ -67,7 +67,7 @@ final class AuditInfoExtractor {
         return auditInfo;
     }
 
-    static Integer
+    static Long
     getResponseLength(Message message) {
         try {
             CachedOutputStream outputStream = (CachedOutputStream) message.getContent(OutputStream.class);
